feat: service로직 user로 변경
