+++ go-assignment/racingcar/cmd/input_test.go
@@ -1,7 +1,8 @@
-package pkg
+package main
 
 import (
 	"github.com/google/go-cmp/cmp"
+	"racing-car/racingcar/pkg/user"
 	"testing"
 )
 
@@ -9,17 +10,17 @@ func Test_inputName(t *testing.T) {
 	data := []struct {
 		testName string
 		name     string
-		expected []string
+		expected []*user.User
 		errMsg   string
 	}{
 		{
 			testName: "정상 테스트",
 			name:     "hunki,hunkis,hunkiss",
-			expected: []string{"hunki", "hunkis", "hunkiss"},
+			expected: []*user.User{{Name: "hunki"}, {Name: "hunkis"}, {Name: "hunkiss"}},
 		},
 		{
 			testName: "경계값 0 테스트",
-			name:     "hunki,,",
+			name:     ",,",
 			errMsg:   "name must be greater than or equal to 1",
 		},
 		{
@@ -36,11 +37,11 @@ func Test_inputName(t *testing.T) {
 
 	for _, d := range data {
 		t.Run(d.testName, func(t *testing.T) {
-			result, err := convertSlice(d.name)
+			users, err := user.CreateUsers(d.name)
 			if err != nil && err.Error() != d.errMsg {
 				t.Errorf("Expected error message %s, got %s", d.errMsg, err.Error())
 			}
-			if diff := cmp.Diff(result, d.expected); diff != "" {
+			if diff := cmp.Diff(users, d.expected); diff != "" {
 				t.Errorf(diff)
 			}
 		})
