+++ go-assignment/racingcar/pkg/user/user_test.go
@@ -1,4 +1,4 @@
-package pkg
+package user
 
 import (
 	"testing"
@@ -7,13 +7,13 @@ import (
 func Test_createUsers(t *testing.T) {
 	data := []struct {
 		testName string
-		names    []string
+		names    string
 		expected []User
 		errMsg   string
 	}{
 		{
 			testName: "정상 테스트",
-			names:    []string{"user1", "user2", "user3"},
+			names:    "이름:user1,user2,user3",
 			expected: []User{
 				{"user1", 0},
 				{"user2", 0},
@@ -22,7 +22,7 @@ func Test_createUsers(t *testing.T) {
 		},
 		{
 			testName: "중복입력 테스트",
-			names:    []string{"hunki", "hunki", "hunkis", "hunkis", "hunkiss"},
+			names:    "이름:hunki,hunki,hunkis,hunkis,hunkiss",
 			expected: []User{
 				{Name: "hunki", NumberOfTurns: 0},
 				{Name: "hunkis", NumberOfTurns: 0},
@@ -33,7 +33,7 @@ func Test_createUsers(t *testing.T) {
 
 	for _, d := range data {
 		t.Run(d.testName, func(t *testing.T) {
-			result := CreateUsers(d.names)
+			result, _ := CreateUsers(d.names)
 			if len(result) != len(d.expected) {
 				t.Error("The number of users created is different")
 			}
